+++ platform/tooling/index.d.ts
@@ -13,4 +13,3 @@ export * from "./src/utils/cost-allocation-tags";
 export * from "./src/utils/config-loader";
 export * from "./src/aspects/log-retention-aspect";
 export * from "./src/nag/nag-packs";
-//# sourceMappingURL=index.d.ts.map